Update mpgRequest toXml() to accept txn type at any position in the txn array
The Moneris developer docs are not explicit about the type needing to be the first value in the transaction array. Since it is already an associative array I propose that the 'type' key be used when determining the transaction type instead of relying on it being the first value in the array provided by the merchant.